Run selected validation tasks on 'Receive SIP'

+++ ESSArch_TA/preingest/tasks.py
@@ -19,7 +19,7 @@ from ESSArch_Core.ip.models import InformationPackage
 from ESSArch_Core.profiles.models import ProfileIP
 from ESSArch_Core.WorkflowEngine.models import ProcessStep, ProcessTask
 
-from ESSArch_Core.util import getSchemas
+from ESSArch_Core.util import getSchemas, get_value_from_path, remove_prefix
 
 class PrepareIP(DBTask):
     event_type = 10100
@@ -152,28 +152,22 @@ class CreatePhysicalModel(DBTask):
 class ReceiveSIP(DBTask):
     event_type = 20100
 
-    def run(self, id=None, label=None, create_date=None, step=None):
-        ip = InformationPackage.objects.create(
-            id=id,
-            Label=label,
-            State="Receiving",
-        )
-        ip.CreateDate = create_date
-        ip.save()
-
-        self.taskobj.information_package = ip
-        self.taskobj.save()
-
-        if step is not None:
-            step.information_package = ip
-            step.save()
-
+    def run(self, ip=None, label=None, create_date=None):
         reception = Path.objects.get(entity="path_ingest_reception").value
         prepare = Path.objects.get(entity="path_ingest_prepare").value
 
-        src = os.path.join(reception, "%s.tar" % ip.pk)
-        dst = os.path.join(prepare, "%s.tar" % ip.pk)
-        shutil.copy(src, dst)
+        tarfile = os.path.join(reception, "%s.tar" % ip.pk)
+        zipfile = os.path.join(reception, "%s.zip" % ip.pk)
+
+        if os.path.isfile(tarfile):
+            src = tarfile
+            dst = os.path.join(prepare, "%s.tar" % ip.pk)
+            shutil.copy(src, dst)
+
+        if os.path.isfile(zipfile):
+            src = zipfile
+            dst = os.path.join(prepare, "%s.zip" % ip.pk)
+            shutil.copy(src, dst)
 
         src = os.path.join(reception, "%s.xml" % ip.pk)
         dst = os.path.join(prepare, "%s.xml" % ip.pk)
@@ -192,7 +186,7 @@ class ReceiveSIP(DBTask):
         self.set_progress(100, total=100)
         return ip
 
-    def undo(self, id=None, label=None, create_date=None, step=None):
+    def undo(self, id=None, label=None, create_date=None):
         pass
 
 class CalculateChecksum(DBTask):
@@ -524,16 +518,11 @@ class CopySchemas(DBTask):
 
 
 class ValidateFiles(DBTask):
-    def run(self, ip, mets_path, validate_fileformat=True, validate_integrity=True):
-        metsdoc = etree.ElementTree(file=mets_path)
 
-        root = metsdoc.getroot()
-        nsmap = {k:v for k,v in root.nsmap.iteritems() if k}
+    def run(self, ip=None, xmlfile=None, validate_fileformat=True, validate_integrity=True, rootdir=""):
+        doc = etree.ElementTree(file=xmlfile)
 
-        prepare_path = Path.objects.get(
-            entity="path_ingest_prepare"
-        ).value
-
+        print "validate files"
         step = ProcessStep.objects.create(
             name="Validate Files",
             parallel=True,
@@ -541,42 +530,43 @@ class ValidateFiles(DBTask):
         )
 
         if any([validate_fileformat, validate_integrity]):
-            for f in metsdoc.findall('.//mets:file', nsmap):
-                filename = f.find('mets:FLocat', nsmap).get('{%s}href' % nsmap['xlink'])
-                filename = os.path.join(prepare_path, filename)
-                try:
-                    fileformat = f.get("{%s}FILEFORMATNAME" % nsmap['ext'])
-                except KeyError:
-                    fileformat = None
-                checksum = f.get("CHECKSUM")
+            for elname, props in settings.FILE_ELEMENTS.iteritems():
+                for f in doc.xpath('.//*[local-name()="%s"]' % elname):
+                    fpath = get_value_from_path(f, props["path"])
 
-                if validate_fileformat and fileformat is not None:
-                    step.tasks.add(ProcessTask.objects.create(
-                        name="preingest.tasks.ValidateFileFormat",
-                        params={
-                            "filename": filename,
-                            "fileformat": fileformat,
-                        },
-                        information_package=ip
-                    ))
+                    if fpath:
+                        fpath = remove_prefix(fpath, props.get("pathprefix", ""))
 
-                if validate_integrity:
-                    step.tasks.add(ProcessTask.objects.create(
-                        name="preingest.tasks.ValidateIntegrity",
-                        params={
-                            "filename": filename,
-                            "checksum": checksum,
-                        },
-                        information_package=ip
-                    ))
+                    fformat = get_value_from_path(f, props.get("format"))
+                    checksum = get_value_from_path(f, props.get("checksum"))
 
+                    if validate_fileformat and fformat is not None:
+                        step.tasks.add(ProcessTask.objects.create(
+                            name="preingest.tasks.ValidateFileFormat",
+                            params={
+                                "filename": os.path.join(rootdir, fpath),
+                                "fileformat": fformat,
+                            },
+                            information_package=ip
+                        ))
+
+                    if validate_integrity and checksum is not None:
+                        step.tasks.add(ProcessTask.objects.create(
+                            name="preingest.tasks.ValidateIntegrity",
+                            params={
+                                "filename": os.path.join(rootdir, fpath),
+                                "checksum": checksum,
+                            },
+                            information_package=ip
+                        ))
 
         self.set_progress(100, total=100)
 
-        step.run()
+        return step.run()
 
-    def undo(self, ip, mets_path, validate_fileformat=True, validate_integrity=True):
+    def undo(self, ip=None, xmlfile=None, validate_fileformat=True, validate_integrity=True, rootdir=""):
         pass
+
 
 class ValidateFileFormat(DBTask):
     event_type = 20260
@@ -632,7 +622,7 @@ class ValidateXMLFile(DBTask):
 
 
 class ValidateLogicalPhysicalRepresentation(DBTask):
-    event_type = 10262
+    event_type = 20262
 
     """
     Validates the logical and physical representation of objects.
@@ -643,46 +633,52 @@ class ValidateLogicalPhysicalRepresentation(DBTask):
     See http://stackoverflow.com/a/7829388/1523238
     """
 
-    def run(self, ip=None, mets_path=None):
-        objpath = ip.ObjectPath
+    def run(self, dirname=None, files=[], xmlfile=None):
+        if dirname:
+            xmlrelpath = os.path.relpath(xmlfile, dirname)
+            xmlrelpath = remove_prefix(xmlrelpath, "./")
+        else:
+            xmlrelpath = xmlfile
 
-        metsdoc = etree.ElementTree(file=mets_path)
+        doc = etree.ElementTree(file=xmlfile)
 
-        root = metsdoc.getroot()
-        nsmap = {k:v for k,v in root.nsmap.iteritems() if k}
+        root = doc.getroot()
 
-        logical_files = []
-        physical_files = []
+        logical_files = set()
+        physical_files = set()
 
-        for f in metsdoc.xpath('.//mets:file | .//mets:mdRef', namespaces=nsmap):
-            if f.tag == "{%s}mdRef" % nsmap['mets']:
-                filename = f.get('{%s}href' % nsmap['xlink'])
-            else:
-                filename = f.find('mets:FLocat', nsmap).get('{%s}href' % nsmap['xlink'])
+        for elname, props in settings.FILE_ELEMENTS.iteritems():
+            for f in doc.xpath('.//*[local-name()="%s"]' % elname):
+                filename = get_value_from_path(f, props["path"])
 
-            if filename:
-                filename = os.path.join(objpath, filename)
-                logical_files.append(filename)
+                if filename:
+                    filename = remove_prefix(filename, props.get("pathprefix", ""))
+                    logical_files.add(filename)
 
-        for root, dirs, files in os.walk(objpath):
-            for f in files:
-                filename = os.path.join(root, f)
+        if dirname:
+            for root, dirs, filenames in os.walk(dirname):
+                for f in filenames:
+                    if f != xmlrelpath:
+                        reldir = os.path.relpath(root, dirname)
+                        relfile = os.path.join(reldir, f)
+                        relfile = remove_prefix(relfile, "./")
 
-                if filename != mets_path:
-                    physical_files.append(filename)
+                        physical_files.add(relfile)
+
+        for f in files:
+            physical_files.add(os.path.basename(f))
 
         print "logical: %s" % logical_files
         print "physical: %s" % physical_files
 
+        assert logical_files == physical_files, "the logical representation differs from the physical"
         self.set_progress(100, total=100)
-        assert Counter(logical_files) == Counter(physical_files), "the physical representation is not valid"
 
-    def undo(self, ip=None, mets_path=None):
+    def undo(self, dirname=None, files=[], xmlfile=None):
         pass
 
-    def get_event_args(self, ip=None, mets_path=None):
-        return [mets_path, ip.ObjectPath]
-
+    def get_event_args(self, dirname=None, files=[], xmlfile=None):
+        return [xmlfile, dirname]
 
 class ValidateIntegrity(DBTask):
     event_type = 20263
